Fix issue 5: PATH not modified
Fix issue #5 (https://github.com/bjoernalbers/aruba-doubles/issues/5)
which reported the PATH was not being updated. The code used ENV['PATH']
directly but was broken when Aruba changed to freeze the environment for
a test.

To fix the issue, the assignments to ENV['PATH'] need to be replaced
with a call to Aruba::Api::Environment::set_environment_variable and
references to ENV['PATH'] need to be replace with references to
Aruba::Runtime::environment['PATH'].

Since the set_environment_variable and environment methods aren't
directly accessible from within ArubaDoubles::Double, the Cucumber
before hook is modified to pass self (i.e., the current World) to the
setup method, which saves it as @world. The methods are then accessible
as @world.set_environment_variable and @world.aruba.environment['PATH'].

Replacing direct access to ENV['PATH'] with the Aruba methods would fix
the Cucumber features but as the RSpec specs don't have a World, it
would break them. So, the old method of accessing ENV['PATH'] must
continue to be supported as well. Add new methods get_path and set_path
to encapsulate access to ENV['PATH']. When RSpec is run, @world will be
nil but when Cucumber is run, @world will be set to the World so
get_path and set_path check @world to decide whether to modify
ENV['PATH'] directly or to use the Aruba methods to modify ENV['PATH'].