Update feature tests
The feature tests create a script and then test whether the script
can be doubled or not. When Aruba runs, it creates a temporary
directory and runs from there. When a file is created (e.g., using the
"Give a file named" matcher), it is created in the bin subdirectory of
the Aruba temporary directory. Aruba always adds the path to that bin
subdirectory as the first entry in the PATH before a command is run.
This means that any script written in a previous step will always be
found first. Aruba does provide the prepend_enviroment variable method
but the strings are prepended before Aruba adds the bin subdirectory to
the path and there appears to be no way to override this without
patching Aruba.

So, modify the features to test using a command which will always be
available on the system (at least any LSB compliant system). This allows
the doubling to be tested, while avoiding the problem Aruba introduces
by forcing the bin subdirectory as the first item in the PATH.